refactoring(Shape): кастомный хук вынесен в отдельный компонент
