+++ CircuitConstructor/CircuitConstructor/ClientApp/src/components/services/useComponentVisible.jsx
@@ -1,0 +1,21 @@
+import { useState, useEffect, useRef } from 'react';
+
+export default function useComponentVisible(initialIsVisible) {
+    const [isComponentVisible, setIsComponentVisible] = useState(false);
+    const ref = useRef(null);
+
+    const handleClickOutside = (event) => {
+        if (ref.current && !ref.current.contains(event.target)) {
+            setIsComponentVisible(false);
+        }
+    };
+
+    useEffect(() => {
+        document.addEventListener('click', handleClickOutside, true);
+        return () => {
+            document.removeEventListener('click', handleClickOutside, true);
+        };
+    }, []);
+
+    return { ref, isComponentVisible, setIsComponentVisible };
+}
